Write input contents on key press
Now instead of printing the current input to the Irssi window, write it
to a script-specific file.

The filename is inspired by Git. I figure we'll probably want to remove
the file when we're done with it. Debated between creating a real temp
file and using this one. We'll decide that later.