FIFO sort of works
After a ton of trial and error, I have something that kind of works.
Couldn't figure out forking for the moment, so this does block Irssi.
And if you do something like change the window layout in Tmux with
<prefix>-M-1, it messes up and closes the FIFO prematurely.

Key points were:

* Opening Vim before making the FIFO, otherwise Vim will block and not
  start.
* Opening Vim with a random other file instead of the FIFO because it
  hasn't been created yet. On write, Vim is set up to write to the FIFO.

Luckily chat messages don't get messed up when the interface blocks, but
I'd like to fix that urgently.