Use two pipe to communicate between forks
Instead of passing messages (prompt input) and commands (location of
temp file) on the same pipe from the child fork to the parent, use
separate pipes. This eliminates blocking behaviour and prints the temp
file location right away. It is then able to read from the socket, and
feed the text up the pipes to be printed too.

Stupid error on my part feeding `@args` to the second `Irssi::input_add`
instead of `@ar2`. Yay, it works! Suck it Perl!