Put the FIFO back in
Now that we have two pipes to communicate between our forks, test the
FIFO again instead of the socket.

The reason why I tried using a socket instead was because I was hoping
that would get rid of the pipe blocking, but it didn't. Instead it
exhibited the same blocking behaviour as the pipe.

Re-enable the FIFO instead of the socket. I'm going to go with this
version instead because it seems like a pipe is more light-weight and it
works basically the same.