Open Vim instance without blocking
I tried to open the Vim instance in `pipe_open_tmux_split`, but this
gave me the same problems as before with Vim not able to write to the
pipe. This was because the main Irssi process was blocked on the `open`
and wasn't reading from the pipe.

The mistake was opening Vim from the parent Irssi. Instead, open it from
the child fork so we don't block the parent and all is good. We can open
Vim and correctly save to the pipe. Yeah!