open_tmux_split: Remove filename first argument
The filename isn't useful to us any more since we can't open the FIFO
directly from Vim (`vim $fifo`). But, in order to be able to close Vim
after writing without it complaining that we haven't written the buffer
to a file, we need to appease it by giving it a random filename.