Remove second pipe
We don't need this any more. It was supposed take the tempfile name from
the child fork and open a Vim instance with that filename, but we found
out we couldn't do that in the parent process, we had to do it in the
fork. So we have no need to pass the tempfile name to the parent process
and no need for this pipe.

Also thankfully gets rid of obscenities.