Send errors from child process to parent and print to Irssi
We previously `die`d in the child process when we encountered an error,
but this didn't kill the child process. Instead, we needed to
`POSIX::_exit(1);` in order to really exit. When `die`ing, the `print`
messages in the child would actally get up to the parent, but now that
we exit correctly they no longer do.

In order to get the messages to the parent so the parent can print them
to Irssi, I tried a couple things. First I tried creating a new pipe
between the parent and child, but this didn't get me the messages
printing.

Asked on Freenode#irssi, and Nei suggested using the same pipe I'm
already using but with a distinguishing prefix for 'ok' and 'error'.
Gave it a try and this time it worked! Yay!