Remove `vimput` Irssi command
It doesn't really make sense to have a `/vimput` command since the
script modifies the input line. It's intended to be run on an existing
input line, not have you enter a command on the prompt.

Instead, change the command to a normal subroutine and instruct users in
the 'help' to create a binding to the Perl function in order to execute
the script.

Thanks to tmux-nicklist-portable for this idea:
https://github.com/irssi/scripts.irssi.org/blob/a0e86203562bac00ecc271842d6f4b238dae1c81/scripts/tmux-nicklist-portable.pl