Fix `/help` completion
When I removed the `/vimput` command in
84ab68f96634cc29eb63a266a30b3147fec835e5, it broke help completion. You
were no longer able to get to our help from:

    /help vi<Tab>

This was because I had removed the command, and Irssi relies on commands
in order to set up completion for `/help`.

To hack completion into `/help`, use Irssi's subcommand feature to
provide a custom `/help vimput` subcommand to `/help`, and have it
display our help as before.