open_tmux_split: Mirror cursor position in Vim
Get the current cursor position in Irssi and make the Vim cursor
position the same. Makes editing more cognitively coherent. This way, if
you've already been editing the Irssi line and the cursor is in the
middle of the line, you can pick up right where you left off in Vim.